# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FieldForce Live — Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions:
    get:
      operationId: listUserSessions
      tags: [history]
      summary: Login/logout session history for a user, filterable by date range
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionRecord"

  # ---------- Movement history / breadcrumb ----------
  /breadcrumb:
    get:
      operationId: getUserBreadcrumb
      tags: [history]
      summary: Location pings for a single day (for map polyline replay)
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Pings for the day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationPing"
  /users/{id}/places-calendar:
    get:
      operationId: getUserPlacesCalendar
      tags: [history]
      summary: 30-day calendar of distinct places visited (derived dwell segments) with total dwell time per place per day
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Calendar days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacesCalendarDay"
  /dwell-segments:
    get:
      operationId: getUserDwellSegments
      tags: [history]
      summary: Dwell segments for a specific day
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Dwell segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DwellSegment"

  # ---------- Day plans / visit stops ----------
...
  /ingest/location:
    post:
      operationId: ingestLocation
      tags: [ingest]
      summary: Ingest one or more location pings (batched for offline catch-up)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestLocationInput"
      responses:
        "201":
          description: Ingested
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationPing"
  /ingest/session:
    post:
      operationId: ingestSession
      tags: [ingest]
      summary: Ingest a login/logout event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestSessionInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRecord"
  /user/dayplan:
    get:
      operationId: getMobileDayPlan
      tags: [ingest]
      summary: Mobile app fetch of the published, priority-ordered day route for a device-authenticated user
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Published day plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayPlanDetail"
        "404":
          description: No published plan for that date

  # ---------- Dispositions (admin CRUD) ----------
...
  /attendance:
    get:
      operationId: getUserAttendanceReport
      tags: [history]
      summary: Attendance records derived from sessions for a user in a date range
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"

  # ---------- Mobile auth ----------
...
components:
  schemas:
...
    VisitStopPriority:
      type: string
      enum: [P1, P2, P3]
    VisitStopStatus:
      type: string
      enum: [PENDING, EN_ROUTE, REACHED, COMPLETED, SKIPPED]
    VisitStopInputType:
      type: string
      enum: [ADDRESS, PIN, LATLNG]
    DayPlanStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    SessionRecord:
      type: object
      required: [id, userId, loginAt, loginLat, loginLng]
      properties:
        id: { type: integer }
        userId: { type: integer }
        loginAt: { type: string, format: date-time }
        loginLat: { type: number }
        loginLng: { type: number }
        logoutAt: { type: ["string", "null"], format: date-time }
        logoutLat: { type: ["number", "null"] }
        logoutLng: { type: ["number", "null"] }
    LocationPing:
      type: object
      required: [id, userId, latitude, longitude, recordedAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        latitude: { type: number }
        longitude: { type: number }
        speedKph: { type: ["number", "null"] }
        accuracyM: { type: ["number", "null"] }
        batteryLevel: { type: ["integer", "null"] }
        recordedAt: { type: string, format: date-time }
    DwellSegment:
      type: object
      required: [id, userId, latitude, longitude, enteredAt, durationSeconds]
      properties:
        id: { type: integer }
        userId: { type: integer }
        latitude: { type: number }
        longitude: { type: number }
        placeLabel: { type: ["string", "null"] }
        enteredAt: { type: string, format: date-time }
        exitedAt: { type: ["string", "null"], format: date-time }
        durationSeconds: { type: integer }
    PlacesCalendarDay:
      type: object
      required: [date, places]
      properties:
        date: { type: string, format: date }
        places:
          type: array
          items:
            $ref: "#/components/schemas/PlacesCalendarEntry"
    PlacesCalendarEntry:
      type: object
      required: [label, totalDwellSeconds]
      properties:
        label: { type: string }
        totalDwellSeconds: { type: integer }
    VisitStop:
      type: object
      required:
        [
          id,
          userId,
          dayPlanId,
          visitDate,
          sequence,
          priority,
          customerCode,
          label,
          inputType,
          rawInput,
          latitude,
          longitude,
          status,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        dayPlanId: { type: integer }
        visitDate: { type: string, format: date }
        sequence: { type: ["integer", "null"] }
        priority:
          $ref: "#/components/schemas/VisitStopPriority"
        customerCode: { type: string }
        label: { type: ["string", "null"] }
        inputType:
          $ref: "#/components/schemas/VisitStopInputType"
        rawInput: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        status:
          $ref: "#/components/schemas/VisitStopStatus"
        plannedArrivalAt: { type: ["string", "null"], format: date-time }
        actualArrivalAt: { type: ["string", "null"], format: date-time }
        contactName: { type: ["string", "null"] }
        contactPhone: { type: ["string", "null"] }
        reachedAt: { type: ["string", "null"], format: date-time }
        startedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        dispositionId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
...
    DayPlanDetail:
      type: object
      required:
        [
          id,
          userId,
          visitDate,
          status,
          totalDistanceMeters,
          totalEtaSeconds,
          stops,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        visitDate: { type: string, format: date }
        status:
          $ref: "#/components/schemas/DayPlanStatus"
        publishedAt: { type: ["string", "null"], format: date-time }
        totalDistanceMeters: { type: number }
        totalEtaSeconds: { type: number }
        stops:
          type: array
          items:
            $ref: "#/components/schemas/VisitStop"
...
    IngestLocationPing:
      type: object
      required: [userId, latitude, longitude, recordedAt]
      properties:
        userId: { type: integer }
        latitude: { type: number }
        longitude: { type: number }
        accuracyM: { type: number }
        batteryLevel: { type: integer }
        recordedAt: { type: string, format: date-time }
    IngestLocationInput:
      type: object
      required: [pings]
      properties:
        pings:
          type: array
          items:
            $ref: "#/components/schemas/IngestLocationPing"
    IngestSessionInput:
      type: object
      required: [userId, event, latitude, longitude, at]
      properties:
        userId: { type: integer }
        event: { type: string, enum: [LOGIN, LOGOUT] }
        latitude: { type: number }
        longitude: { type: number }
        at: { type: string, format: date-time }
...
    AttendanceRecord:
      type: object
      required: [date, loginAt, loginLat, loginLng]
      properties:
        date: { type: string, format: date }
        loginAt: { type: string, format: date-time }
        loginLat: { type: number }
        loginLng: { type: number }
        logoutAt: { type: ["string", "null"], format: date-time }
        logoutLat: { type: ["number", "null"] }
        logoutLng: { type: ["number", "null"] }
        totalHours: { type: ["number", "null"] }
...